refactor: improve portability of timeval string formatting
Cast tv_sec to long long and tv_usec to long to ensure consistent
behavior across 32-bit and 64-bit architectures and avoid format
string warnings/mismatches.

+++ lib/nsutils.c
@@ -70,7 +70,7 @@ float tv_delta_f(const struct timeval *start, const struct timeval *stop)
 
 /* format duration seconds into human readable string */
 const char* tv_str(struct timeval *tv) {
-	return (char *)mkstr("%lu.%06lu", tv->tv_sec, tv->tv_usec);
+	return (char *)mkstr("%lld.%06ld", (long long)tv->tv_sec, (long)tv->tv_usec);
 }
 
  /* Convert string to timeval */
